% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citoInput.R
\name{citoInput}
\alias{citoInput}
\title{Convert generated data in R to cito text files - needs further dev}
\usage{
citoInput(
  adat,
  nsegments = 1,
  minlen = 10,
  maxlen = minlen,
  pval = 0.6,
  ntests = 1,
  scale = 1,
  folder = file.path(getwd(), "cito/seeding")
)
}
\arguments{
\item{folder}{}
}
\value{

}
\description{
Convert generated data in R to cito text files - needs further dev
}
