% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoselect.R
\name{selectItem}
\alias{selectItem}
\title{Title}
\usage{
selectItem(
  scalename,
  ability,
  items = msAlgo::items,
  targetease = 0.1,
  samplesize = 1
)
}
\arguments{
\item{samplesize}{}
}
\value{

}
\description{
Title
}
